refactor: separate inbount, outbound structure in cluster
