+++ src/services/cluster/inbound/request.rs
@@ -1,4 +1,4 @@
-use super::arguments::QueryArguments;
+use crate::services::cluster::inbound::arguments::QueryArguments;
 use crate::services::query_io::QueryIO;
 use anyhow::Context;
 
@@ -40,10 +40,10 @@ impl HandShakeRequest {
 
         match self.args.as_mut_slice() {
             [QueryIO::BulkString(key), QueryIO::BulkString(port), ..]
-                if key == "listening-port" =>
-            {
-                Ok(port.parse::<u16>()?)
-            }
+            if key == "listening-port" =>
+                {
+                    Ok(port.parse::<u16>()?)
+                }
             _ => Err(anyhow::anyhow!("Invalid listening-port arguments")),
         }
     }
