+++ src/services/cluster/mod.rs
@@ -1,6 +1,5 @@
 pub mod actor;
 mod command;
-pub mod establishment;
-pub mod inbound_stream;
 pub mod manager;
-pub mod outbound_stream;
+pub mod inbound;
+pub mod outbound;
